V108 Incorrect index type: m_connections[not a memsize-type]. Use memsize type instead. The analyzer found a possible error of indexing large arrays.
V104 Implicit conversion of 'i' to memsize type in an arithmetic expression: m_connections.begin() + i

+++ dev/Gems/GraphCanvas/Code/Source/Components/Slots/Data/DataSlotComponent.cpp
@@ -375,7 +375,7 @@ namespace GraphCanvas
 
             if (m_displayedConnection.IsValid())
             {
-                for (int i = static_cast<int>(m_connections.size()) - 1; i >= 0; ++i)
+                for (size_t i = m_connections.size() - 1; i >= 0; ++i)
                 {
                     if (m_connections[i] == m_displayedConnection)
                     {
